Remove aws/credential_utils.py to extractors parent package
