+++ nodestream/pipeline/extractors/credential_utils.py
