+++ nodestream/pipeline/extractors/stores/aws/athena_extractor.py
@@ -4,7 +4,7 @@ from logging import getLogger
 from typing import Any, AsyncGenerator
 
 from ...extractor import Extractor
-from .credential_utils import AwsClientFactory
+from ...credential_utils import AwsClientFactory
 
 ATHENA_STATE_RUNNING = "RUNNING"
 ATHENA_STATE_FAILED = "FAILED"
